Refactor mock setup functions to return spies
Jira ticket: CAMS-368
